# backend/app/api/static/docs/swagger.yaml
basePath: /api
definitions:
...
  repo.DocumentOut:
    properties:
      id:
        type: string
      path:
        type: string
      title:
        type: string
    type: object
...
  repo.ItemAttachment:
    properties:
      createdAt:
        type: string
      document:
        $ref: '#/definitions/repo.DocumentOut'
      id:
        type: string
      primary:
        type: boolean
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  repo.ItemCreate:
    properties:
      description:
        maxLength: 1000
        type: string
      labelIds:
        items:
          type: string
        type: array
      locationId:
        description: Edges
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      parentId:
        type: string
        x-nullable: true
    required:
    - name
    type: object
  repo.ItemField:
    properties:
      booleanValue:
        type: boolean
      id:
        type: string
      name:
        type: string
      numberValue:
        type: integer
      textValue:
        type: string
      type:
        type: string
    type: object
  repo.ItemOut:
    properties:
      archived:
        type: boolean
      assetId:
        example: "0"
        type: string
      attachments:
        items:
          $ref: '#/definitions/repo.ItemAttachment'
        type: array
      createdAt:
        type: string
      description:
        type: string
      fields:
        items:
          $ref: '#/definitions/repo.ItemField'
        type: array
      id:
        type: string
      imageId:
        type: string
      insured:
        type: boolean
      labels:
        items:
          $ref: '#/definitions/repo.LabelSummary'
        type: array
      lifetimeWarranty:
        description: Warranty
        type: boolean
      location:
        allOf:
        - $ref: '#/definitions/repo.LocationSummary'
        description: Edges
        x-nullable: true
        x-omitempty: true
      manufacturer:
        type: string
      modelNumber:
        type: string
      name:
        type: string
      notes:
        description: Extras
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/repo.ItemSummary'
        x-nullable: true
        x-omitempty: true
      purchaseFrom:
        type: string
      purchasePrice:
        type: number
      purchaseTime:
        description: Purchase
        type: string
      quantity:
        type: integer
      serialNumber:
        type: string
      soldNotes:
        type: string
      soldPrice:
        type: number
      soldTime:
        description: Sold
        type: string
      soldTo:
        type: string
      syncChildItemsLocations:
        type: boolean
      updatedAt:
        type: string
      warrantyDetails:
        type: string
      warrantyExpires:
        type: string
    type: object
  repo.ItemPatch:
    properties:
      id:
        type: string
      quantity:
        type: integer
        x-nullable: true
        x-omitempty: true
    type: object
  repo.ItemPath:
    properties:
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/repo.ItemType'
    type: object
  repo.ItemSummary:
    properties:
      archived:
        type: boolean
      assetId:
        example: "0"
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      imageId:
        type: string
      insured:
        type: boolean
      labels:
        items:
          $ref: '#/definitions/repo.LabelSummary'
        type: array
      location:
        allOf:
        - $ref: '#/definitions/repo.LocationSummary'
        description: Edges
        x-nullable: true
        x-omitempty: true
      name:
        type: string
      purchasePrice:
        type: number
      quantity:
        type: integer
      updatedAt:
        type: string
    type: object
  repo.ItemType:
    enum:
    - location
    - item
    type: string
    x-enum-varnames:
    - ItemTypeLocation
    - ItemTypeItem
  repo.ItemUpdate:
    properties:
      archived:
        type: boolean
      assetId:
        type: string
      description:
        maxLength: 1000
        type: string
      fields:
        items:
          $ref: '#/definitions/repo.ItemField'
        type: array
      id:
        type: string
      insured:
        type: boolean
      labelIds:
        items:
          type: string
        type: array
      lifetimeWarranty:
        description: Warranty
        type: boolean
      locationId:
        description: Edges
        type: string
      manufacturer:
        type: string
      modelNumber:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      notes:
        description: Extras
        type: string
      parentId:
        type: string
        x-nullable: true
        x-omitempty: true
      purchaseFrom:
        maxLength: 255
        type: string
      purchasePrice:
        type: number
        x-nullable: true
        x-omitempty: true
      purchaseTime:
        description: Purchase
        type: string
      quantity:
        type: integer
      serialNumber:
        description: Identifications
        type: string
      soldNotes:
        type: string
      soldPrice:
        type: number
        x-nullable: true
        x-omitempty: true
      soldTime:
        description: Sold
        type: string
      soldTo:
        maxLength: 255
        type: string
      syncChildItemsLocations:
        type: boolean
      warrantyDetails:
        type: string
      warrantyExpires:
        type: string
    required:
    - name
    type: object
...
  repo.LabelSummary:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  repo.LocationSummary:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  repo.PaginationResult-repo_ItemSummary:
    properties:
      items:
        items:
          $ref: '#/definitions/repo.ItemSummary'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact:
    name: Don't
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  version: "1.0"
paths:
...
  /v1/items:
    get:
      parameters:
      - description: search string
        in: query
        name: q
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: items per page
        in: query
        name: pageSize
        type: integer
      - collectionFormat: multi
        description: label Ids
        in: query
        items:
          type: string
        name: labels
        type: array
      - collectionFormat: multi
        description: location Ids
        in: query
        items:
          type: string
        name: locations
        type: array
      - collectionFormat: multi
        description: parent Ids
        in: query
        items:
          type: string
        name: parentIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.PaginationResult-repo_ItemSummary'
      security:
      - Bearer: []
      summary: Query All Items
      tags:
      - Items
    post:
      parameters:
      - description: Item Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.ItemCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.ItemSummary'
      security:
      - Bearer: []
      summary: Create Item
      tags:
      - Items
  /v1/items/{id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Item
      tags:
      - Items
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ItemOut'
      security:
      - Bearer: []
      summary: Get Item
      tags:
      - Items
    patch:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Item Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.ItemPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ItemOut'
      security:
      - Bearer: []
      summary: Update Item
      tags:
      - Items
    put:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Item Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.ItemUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ItemOut'
      security:
      - Bearer: []
      summary: Update Item
      tags:
      - Items
...
  /v1/items/{id}/path:
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.ItemPath'
            type: array
      security:
      - Bearer: []
      summary: Get the full path of an item
      tags:
      - Items
  /v1/items/export:
    get:
      produces:
      - text/tsv
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - Bearer: []
      summary: Export Items
      tags:
      - Items
  /v1/items/fields:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - Bearer: []
      summary: Get All Custom Field Names
      tags:
      - Items
  /v1/items/fields/values:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - Bearer: []
      summary: Get All Custom Field Values
      tags:
      - Items
  /v1/items/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image to upload
        in: formData
        name: csv
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Import Items
      tags:
      - Items
...
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"